+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ## Versions
 
+- [0.3.0](#030)
 - [0.2.0](#020)
 - [0.1.0](#010)
 
@@ -10,6 +11,15 @@
 ---
 
 ## Changes
+
+<br>
+
+### 0.3.0
+
+- `basename` included into every `routes` route during init process
+- `getAppBasename` selector removed
+- Cleanup
+
 
 <br>